// src/app/Models/Lascellen/RobotProgrammaCelA003.ts
import { RobotProgrammaMachineOnderdeelProduct } from "./RobotProgrammaMachineOnderdeelProduct";
import { RobotProgramma } from "./RobotProgramma";
export class RobotProgrammaCelA003 extends RobotProgramma {
  malLinks: RobotProgrammaMachineOnderdeelProduct;
  malRechts: RobotProgrammaMachineOnderdeelProduct;
  malCompleet: RobotProgrammaMachineOnderdeelProduct;
}